+++ HISTORY.md
@@ -7,6 +7,10 @@
 * **Notable Fixes:**
     * SciSnap file reader, thanks, Eckart!
 
+### 2025-01-06
+* threads: optimized rapid evaluation of generic conditions & custom hat blocks
+* objects, threads: harmonized Quickstep / Doublestep interaction
+
 ### 2025-01-04
 * objects: enabled "quicksteps" to capture atomic custom events
 * updated version